backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.20
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
openai==1.54.3
python-dotenv==1.0.1
pypdf==5.3.1
python-pptx==1.0.2
python-docx==1.1.2
pytesseract==0.3.13
